chat_app_fastapi/requirements.txt
aio-pika==9.5.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiormq==6.8.1
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
attrs==24.2.0
bidict==0.23.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.5
frozenlist==1.5.0
h11==0.14.0
idna==3.10
kombu==5.4.2
multidict==6.1.0
pamqp==3.3.0
propcache==0.2.0
pydantic==2.10.1
pydantic_core==2.27.1
python-engineio==4.10.1
python-socketio==5.11.4
requests==2.32.3
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
vine==5.1.0
wsproto==1.2.0
yarl==1.18.0
